rename: RatedStar 폴더 변경 및 사용컴포넌트에서 경로변경 (#48)
