+++ src/components/basicComponents/ratedStar/RatedStar.tsx
